Property Type not assigned when <10 rows of data
There is a loop that checks the first 10 rows of data when giving a
preview. If there were not 10 rows, the previous check was for ""
instead of null. This was causing an error because it was not skipped
properly. The error prevented the type from properly being updated. This
has been corrected